Well it works now, still need to change some stuff but nothing big
You can add stuff if you want matt.
making stuff with json now works by passing the chunk of json data along
with the index of current section, and all the property getting is done
in the class itself. Btw, for things like getting an animation by a
specific name, pass the json reader made in the game class to the class
that needs it instead of just making a new one.